Move ASTGen source-file handling to its own (Swift) file
